requirements.txt
backports.zoneinfo==0.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
clickhouse-driver==0.2.9
idna==3.11
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.3
